pyproject.toml
[project]
name = "toutiao-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.3.2",
    "cryptography>=46.0.7",
    "fastapi>=0.135.3",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "pydantic>=2.13.1",
    "sqlalchemy[asyncio]>=2.0.49",
    "starlette>=1.0.0",
    "uvicorn>=0.44.0",
    "black>=26.3.1",
    "python-dotenv>=1.2.2",
]
